% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-idx.R
\name{autoplot}
\alias{autoplot}
\title{Autoplot function for calculated indexes}
\usage{
autoplot(data, time = NULL, site = NULL)
}
\arguments{
\item{data}{A {.field idx_res} object from {.code calculate_idx()}.}

\item{time}{the time variable in the data, support non-standard evaluation}

\item{site}{the time variable in the data, support non-standard evaluation}
}
\description{
The function creates a line plot of the indexes calculated from
\code{calculate_idx()}
}
\examples{
\dontrun{
library(dplyr)
library(sf)
d <- disaster |> filter(cat_name == "CAT131") |> disaster_as_shape()
s <- query_species(d[1,])
library(tsibble)
time <- s$occurrence |> pull(eventDate) |> yearmonth() |> unique()
occur <- purrr::map(time, function(x){
  s$occurrence |>
    mutate(ym = yearmonth(eventDate)) |>
    filter(between(ym, x-3, x + 3))
})
names(occur) <- time
res <- occur |>
  purrr::map_dfr(~calculate_idx(occurrence = .x, taxonomy = s$taxonomy), .id = "ym") |>
  mutate(ym = time)
res |> autoplot(time = ym)
}
}
