% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-indexes.R
\name{calculate_idx}
\alias{calculate_idx}
\title{Calculate indexes from the biodiversity data}
\usage{
calculate_idx(
  data = NULL,
  occurrence = NULL,
  taxonomy = NULL,
  is_count = FALSE,
  indexes = c("shannon", "simpson", "CW")
)
}
\arguments{
\item{data}{A \code{galah_res} object from  \code{query_species()}.}

\item{occurrence}{A dataframe of animal occurrence data, which can be
individual occurrence count or occurrence count per species.}

\item{taxonomy}{A dataframe of animal taxonmy data.}

\item{is_count}{Indicator whether the animal occurrence data is recorded
as occurrence count per species.}

\item{indexes}{The index(es) to compute: "shannon", "simpson" or "CW".}
}
\description{
The function calculates biodiversity indexes (Simpson's Diversity Index,
Shannon Diversity Index, and Clarke & Warwick's Distinctness and Diversity
Index) from the animal data queried from the Atlas of Living Australia
}
\examples{
\dontrun{
library(dplyr)
library(sf)
d <- disaster |> filter(cat_name == "CAT131") |> disaster_as_shape()
s <- query_species(d[1,])
s |> calculate_idx()
}
}
\references{
https://www.ala.org.au/
}
