% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disaster}
\alias{disaster}
\alias{test_ala}
\title{Disaster data with postcode information}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6161 rows and 24 columns.

An object of class \code{galah_res} (inherits from \code{list}) of length 2.
}
\usage{
disaster

test_ala
}
\description{
This is the dataset of disasters which comes from Insurance Council of
Australia.  Disaster data without postcode information are removed.
The shape files of postcodes come from the Australian Bureau of Statistics

This dataset contains the occurrence and taxonomy data used in ...
}
\details{
\describe{
\item{column name}{description}
\item{column name}{description}
\item{column name}{description}
\item{column name}{description}
\item{column name}{description}
\item{column name}{description}
}

\describe{
\item{column name}{description}
\item{column name}{description}
\item{column name}{description}
\item{column name}{description}
\item{column name}{description}
\item{column name}{description}
}
}
\references{
https://insurancecouncil.com.au/industry-members/data-hub/
}
\keyword{datasets}
