% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaster-as-shape.R
\name{disaster_as_shape}
\alias{disaster_as_shape}
\title{Find the shapefile for the postcode in the disaster data}
\usage{
disaster_as_shape(disaster_df, postcode = "postcode", as_sfc = FALSE)
}
\arguments{
\item{disaster_df}{A dataframe containing the disaster data}

\item{postcode}{The column name of the postcode in the disaster data}

\item{as_sfc}{Whether to return a simple feature column (sfc) object
(with default as TRUE) or a simple feature (sf) object}
}
\description{
Find the shapefile for the postcode in the disaster data
}
\examples{
\dontrun{
library(dplyr)
disaster |> filter(cat_name == "CAT131") |> disaster_as_shape()
}
}
