% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_time.R
\name{split_time}
\alias{split_time}
\title{Split the occurrence data by month}
\usage{
split_time(data, window = 3)
}
\arguments{
\item{data}{a {.field galah_res} object from {.code query_species}}

\item{window}{the time window to split the data}
}
\description{
Split the occurrence data by month
}
\examples{
res <- test_ala |> split_time()
res[[1]]
}
