% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-indexes.R
\name{calculate_idx}
\alias{calculate_idx}
\title{Calculate indexes from the biodiversity data}
\usage{
calculate_idx(
  data = NULL,
  occurrence = NULL,
  taxonomy = NULL,
  is_count = FALSE,
  indexes = c("shannon", "simpson", "CW")
)
}
\arguments{
\item{data}{A {.field galah_res} object from {.code query_species()}.}

\item{occurrence, }{required if data is not supplied. A data frame of
occurrence data. Can be counts or observation}

\item{taxonomy, }{required for CW indexes are and if data is not supplied .
A data frame of taxonomy data.}

\item{is_count}{whether the occurrence data is count data or observation data.}

\item{indexes}{the indexes to compute: shannon, simpson or CW indexes.}
}
\description{
The function calculates biodiversity indexes (Shannon, Simpson, and
Clarke & Warwick's  diversity and distinctness) from the occurrence and
taxonomy data.
}
\examples{
\dontrun{
library(dplyr)
library(sf)
d <- disaster |> filter(cat_name == "CAT131") |> disaster_as_shape()
s <- query_species(d[1,])
s |> calculate_idx()
}
}
