% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disaster}
\alias{disaster}
\title{Disaster and postcode data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6161 rows and 24 columns.
}
\usage{
disaster
}
\description{
This dataset contains the postcode and disaster data used in ...
}
\details{
Disaster data comes from \link{organisation name}
\url{https://insurancecouncil.com.au/industry-members/data-hub/}. Disaster
without postcode information is removed.

The postcode data comes from ABS ...

\describe{
\item{column name}{description}
\item{column name}{description}
\item{column name}{description}
\item{column name}{description}
\item{column name}{description}
\item{column name}{description}
}
}
\keyword{datasets}
