% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_species.R
\name{query_species}
\alias{query_species}
\title{Query species from the Atlas of Living Australia}
\usage{
query_species(shape, year_vec = c(2019, 2020))
}
\arguments{
\item{shape}{an sfc object}

\item{year_vec}{a vector of years}
}
\description{
The function queries species from the Atlas of Living Australia ...
}
\details{
instruction on need to config with email
}
\examples{
\dontrun{
library(dplyr)
library(sf)
d <- disaster |> filter(cat_name == "CAT131") |> disaster_as_shape()
query_species(d[1,])
}
}
